testing_xuexi/requirements.txt
pytest==8.3.5
pytest-xdist==3.6.1
requests==2.32.3
allure-pytest==2.13.5
pydantic==2.10.6
pymysql==1.1.1
redis==5.2.1
python-dotenv==1.0.1
locust==2.32.3
prometheus-client==0.21.1
cryptography==44.0.2
